[Feat] 유저 profile 기능 구현 - 학점 및 개인 스펙 CRUD

+++ Hstep/src/main/java/com/Hstep/Hstep/domain/track/entity/Track.java
@@ -24,7 +24,7 @@ public class Track {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "track_id")
-    private Integer trackId;
+    private Long trackId;
 
     @Column(name = "track_code", nullable = false, length = 100)
     private String trackCode;
